Deprecated NavigationHelper and most relatec classes (and taglibs), deprecated the vwebroot and url.db subpackages
