+++ gsf-wra-legacy/src/main/java/com/fatwire/gst/foundation/vwebroot/VirtualWebroot.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2012 Oracle Corporation. All Rights Reserved.
+ * Copyright 2010 FatWire Corporation. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,17 +13,27 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package com.fatwire.gst.foundation.vwebroot;
 
-import com.fatwire.gst.foundation.controller.action.Action;
-import com.fatwire.gst.foundation.mobile.action.DeviceAwareRenderPageAction 
+import com.fatwire.assetapi.data.AssetId;
 
-/*
- * Action that just extends the DeviceAwareRenderPageAction. TYhe DeviceAwareRenderPageAction also handles translation of the requested asset.
- * The action is defined in root.groovy. To make this the default action it is easiest to set a default argument action=root at the GSF/Dispatcher SiteCatalog entry.
+/**
+ * Virtual webroot as defined in GST Site Foundation document
+ * 
+ * @author Tony Field
+ * @since Jul 22, 2010
+ * 
+ * @deprecated June 15, 2016 by fvillalba
  * 
  */
+@Deprecated
+public interface VirtualWebroot {
 
-public class RootAction extends DeviceAwareRenderPageAction {
+    public AssetId getId();
 
+    public String getMasterVirtualWebroot();
+
+    public String getEnvironmentVirtualWebroot();
+
+    public String getEnvironmentName();
 }
-
